Add a basic layout for improved table granular reset

+++ CTFd/admin/__init__.py
@@ -14,10 +14,13 @@ from flask import (
     url_for,
 )
 
-from CTFd.cache import cache, clear_config
+from CTFd.cache import cache, clear_config, clear_standings, clear_pages
 from CTFd.models import (
     Awards,
+    Challenges,
     Configs,
+    Notifications,
+    Pages,
     Solves,
     Submissions,
     Teams,
@@ -176,19 +179,52 @@ def config():
 @admins_only
 def reset():
     if request.method == "POST":
-        # Truncate Users, Teams, Submissions, Solves, Notifications, Awards, Unlocks, Tracking
-        Tracking.query.delete()
-        Solves.query.delete()
-        Submissions.query.delete()
-        Awards.query.delete()
-        Unlocks.query.delete()
-        Users.query.delete()
-        Teams.query.delete()
-        set_config("setup", False)
+        require_setup = False
+        logout = False
+        next_url = url_for("admin.statistics")
+
+        data = request.form
+
+        if data.get("pages"):
+            Pages.query.delete()
+            # PageFiles
+
+        if data.get("notifications"):
+            Notifications.query.delete()
+
+        if data.get("challenges"):
+            Challenges.query.delete()
+            # ChallengeFiles
+
+        if data.get("accounts"):
+            Users.query.delete()
+            Teams.query.delete()
+            require_setup = True
+            logout = True
+
+        if data.get("submissions"):
+            Solves.query.delete()
+            Submissions.query.delete()
+            Awards.query.delete()
+            Unlocks.query.delete()
+
+        if require_setup:
+            set_config("setup", False)
+            cache.clear()
+            logout_user()
+            next_url = url_for("views.setup")
+
         db.session.commit()
-        cache.clear()
-        logout_user()
+
+        clear_pages()
+        clear_standings()
+        clear_config()
+
+        if logout is True:
+            cache.clear()
+            logout_user()
+
         db.session.close()
-        return redirect(url_for("views.setup"))
+        return redirect(next_url)
 
     return render_template("admin/reset.html")
